Add more test for JSONIFY
Allows me to find some errors: specifically a from_json method should throws
an exception in case of error, as it is used by the std::variant
deserialization.

Also note that std::variant<std::string, Enum> will fail, as the current
code is not able to check if the enum text is right.

Change-Id: I8c92e8fee5c8625762e1ee04999e55ade4ba3d7f
Reviewed-on: https://codereview.kdab.com/c/knut/+/95892
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Andras Mantia <andras@kdab.com>