Move to QTest for testing
doctest is nice, unfortunately it's lacking support in IDEs, which makes
handling tests harder. Also missing some features like benchmark and
data-driven tests/
The build time penalty is limited, and may be even less by moving to
dynamic libraries (something to research).

The global knut test is commented out for now, as it was a bad idea
(impossible to debug), another commit will fix that.

Change-Id: Ia2db9bc83f34c395893957d7fecfaedd14d26218
Reviewed-on: https://codereview.kdab.com/c/knut/+/97490
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Nicolas Arnaud-Cormos <nicolas@kdab.com>