CMake: Rebuild the documentation on compilation
With this change, CMake will always automatically run cpp2doc when
building.
That way we can make sure the docs/ folder is always up-to-date.

Open Question: On my machine, cpp2doc reorders some of the documentation,
does this only occur because the docs/ folder wasn't up-to-date on my
local machine, or does this also happen for everyone else? This would
then lead to useless merge conflicts :/

Change-Id: I519b98f892cb571a9cbeae66c3a5c43ed78ed8d9
Reviewed-on: https://codereview.kdab.com/c/knut/+/123010
Reviewed-by: Nicolas Arnaud-Cormos <nicolas@kdab.com>
Tested-by: Continuous Integration <build@kdab.com>