Remove JSON API
The JSON API has proven to be a weird middle-ground.
JSON is useful to express a list of transformations, but ultimately
doesn't have the ability to define dependencies and control flow between
transformations (i.e. only run this transformation if you found a certain
function, etc.).
And once we start building stuff like this into JSON, we basically build
our on DSL. As this isn't really our goal, we believe scripting
using Javascript to be the better way forward.
Especially in combination with treesitter queries, this is much more
powerful, flexible and expressive.

Therefore remove the JSON API, before too many people become dependent
upon it.

Change-Id: I0831550738dd242a821f1fd489a5314e5c2992a6
Task-Id: KNUT-77
Reviewed-on: https://codereview.kdab.com/c/knut/+/125612
Reviewed-by: Leon Matthes <leon.matthes@kdab.com>