Add a model of scripts in the current path
This data is currently displayed in the OptionsDialog.
It should make it easier for users to identify whether they added all
necessary script paths.
The model also serves as a foundation for the list of suggested scripts.
Such a list can be created by adding a SortFilterProxyModel to this
model that only shows scripts that can be executed in the current
context.

Task-Id: KNUT-99
Change-Id: I2ae430ef01957d5f19a6f340d4efb406ebc5e241
Reviewed-on: https://codereview.kdab.com/c/knut/+/131029
Reviewed-by: Leon Matthes <leon.matthes@kdab.com>