ScriptSuggestions model & view
In many cases, the arguments to a script can be inferred from the
current document and cursor position.
To accomodate this automatically, the ScriptSuggestions model
infers the currently runnable scripts using TreeSitter queries.
When double-clicking, the script is run with the given context.
Scripts can then decide themselves what to do with the context.

Task-Id: KNUT-99
Change-Id: I9514b78556267a2ad6aedd956e1cd9285088967f
Reviewed-on: https://codereview.kdab.com/c/knut/+/131402
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Nicolas Arnaud-Cormos <nicolas@kdab.com>