Fix some low-hanging errors.
Two of these are memory leaks reported by lsan.
The LogCounter change improves the debugging experience when
trying to evaluate why a LogCount is not as expected.
Furthermore, a bug in the LogCounter is fixed where the logger sink
wasn't disconnected correctly.

Additionally, spdlog is instructed to accept setting the log level
through the SPDLOG_LEVEL environment variable.

Change-Id: I1a97dac1fb9b93d6e589d648bfbc06c4e6bf9798
Reviewed-on: https://codereview.kdab.com/c/knut/+/132088
Reviewed-by: Nicolas Arnaud-Cormos <nicolas@kdab.com>
Tested-by: Continuous Integration <build@kdab.com>