Improve engine cleanup
The engine is now deallocated in more cases.

Allocating a QQmlEngine still leaks around 8KB of data.
Heaptrack indicates this may be due to type registration.
However, this is reused on subsequent runs, so we're not leaking more
than 8KB, which is fine.

The QQuickView window also wasn't deallocated before, which caused 3-4KB
of leaks for every script run!
This is the much larger issue, as that would have continued to increase
with every script run.

This patch fixes this by parenting the QQuickView to the QML engine,
which deletes itself.

Change-Id: I3d4eaa9faf584fcc42dc402800a8b583d4a38dd6
Reviewed-on: https://codereview.kdab.com/c/knut/+/132202
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Nicolas Arnaud-Cormos <nicolas@kdab.com>