Remove all MFC scripts from this repository
Those scripts are now moved to the mfc-utils directory, where they are
kept in sync with the KDWinUtils library.
I left the mfc-convert-dialog (added deprecated to the name) here, as we
still have an integration test for this one in Knut, goal at term is to
move them to mfc-utils, and use this repository as a git submodule of
knut.

Change-Id: Ib672a653f839124e52592c065054a59d78217e38
Reviewed-on: https://codereview.kdab.com/c/knut/+/138614
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Leon Matthes <leon.matthes@kdab.com>
Reviewed-by: Nicolas Arnaud-Cormos <nicolas@kdab.com>