ci: Use spdlogs bundled fmt library
For now, drop support for std::format.
It's not supported by the compilers used in our CI on ubuntu & macOS

At least on my local machine, this means we need to disable -Werror, as
one of the templates in the bundled format library produces a
-Werror=dangling-reference.
We should try to reenable this when possible.