fix: save settings file in different directory in testing mode
before this commit, we didn't save json files in test mode, so as not to overwrite the json file already present in the user's home directory.
now we save this json file in the temp directory.

We have also improved tst_settings.cpp, which now saves the json file in the temp directory.