fix: improve script performance by changing QTreeView settings
This patch determine the height of a row in history panel only once at the beginning and does not resize the width of the first column for each new log. Column resizing will now be done manually by the user to gain performance during execution.
For example, for the mfc_update_translations_qml script:
 - before the patch: ≈80000ms
 - after the patch: ≈500ms