refactor!: indentationAt API now returns int
The `indentationAtLine` and `...AtPosition` methods now return `int`, in
accordance to the `indent` and `setIndentation` methods.

The old behavior is still accessible via the `indentTextAtLine` and
`indentTextAtPosition` functions.
It turned out to still be useful, as there are a few situation where you
just want to copy the indentation from another line.
Getting access to the original indentation is useful there, as it can
deal better with cases where Knuts indentation settings don't match up
with the actual indentation in the file.