feat: automatically show and hide close buttons from tab bar
This implementation shows and hides all close buttons simultaneously
while mouse cursor hovers over them. A delay was added before close
buttons are hidden again to save the user from a jarring experience
when they off-shoot the mouse cursor reaching for the tab bar.

Ideally we'd only display close buttons for the active tab and the tab
currently being hovered. However, the list of tabs is stored in the
protected side of QTabBar (qtabbar_p.h) and we can't make use of it
without re-inventing the wheel.