feat: Use an optimized initial size for translation related columns
Make better use of available width when dealing with Ts translation
files by setting an initial size based on the space available when the
file is opened.

I used the window width as the reference size because the tableView's
width hasn't yet been set at this point in initialization. By
dividing available space by 4, we ensure a reasonable amount of space
is available for the removable sidebar.